+++ src/components/BasicQuestion/QuestionInfo.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {THEME_COLOR} from '../../common/common';
+import {THEME_COLOR} from '../../utils/helper';
 
 const QuestionInfo = ({
   currQuestion,
